refactor(runtime): split system.rs into cohesive modules (§7-8)
Split the 2016-line system.rs (over the 500-line limit) into six
cohesive modules, all under 500 lines:

- system.rs (353)               - CORE_TYPE_NAMES, eval-method injection,
                                  BEGIN phasers, redeclaration scan, suggest-*
- system_eval_redecl.rs (267)   - check_eval_class_redeclarations
- system_eval_vars.rs (381)     - undeclared-variable / placeholder checks
- system_eval_names.rs (371)    - forward-call / undeclared-name checks
- system_eval_string.rs (451)   - eval_eval_string + operator/sub collection
                                  + code-rewrite parse helpers
- system_introspect.rs (211)    - callframe/lexical introspection + host/signal

All methods are impl Interpreter, so the split is visibility-only. A few
cross-module private helpers (parse_and_eval_with_operators,
find_our_routine_redeclaration, suggest_declared_vars, collect_declared_vars,
eval_result_is_unresolved_bareword) are bumped to pub(super). The fn
multiset is byte-identical. No behavior change.
