+++ src/runtime/mod.rs
@@ -259,6 +259,11 @@ mod supply_classify;
 mod supply_promise;
 mod supply_transform;
 mod system;
+mod system_eval_names;
+mod system_eval_redecl;
+mod system_eval_string;
+mod system_eval_vars;
+mod system_introspect;
 mod tap_state;
 mod test_functions;
 pub(crate) mod types;
