+++ src/runtime/system.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::symbol::Symbol;
 
 /// Core built-in type names used as candidates for type-name "Did you mean"
 /// suggestions (in addition to user-registered classes).
@@ -70,7 +69,7 @@ pub(crate) const CORE_TYPE_NAMES: &[&str] = &[
 ];
 
 impl Interpreter {
-    fn parse_and_eval_with_operators(
+    pub(super) fn parse_and_eval_with_operators(
         &mut self,
         src: &str,
         op_names: &[String],
@@ -232,7 +231,7 @@ impl Interpreter {
     /// `our sub foo` declarations install the same package symbol, so a duplicate
     /// across sibling blocks (or block vs mainline) is X::Redeclaration. `my sub`
     /// is lexical and does not conflict across blocks, so it is ignored here.
-    fn find_our_routine_redeclaration(
+    pub(super) fn find_our_routine_redeclaration(
         stmts: &[Stmt],
         seen: &mut HashSet<String>,
     ) -> Option<RuntimeError> {
@@ -267,416 +266,12 @@ impl Interpreter {
         None
     }
 
-    pub(crate) fn check_eval_class_redeclarations(
-        &self,
-        stmts: &[Stmt],
-    ) -> Result<(), RuntimeError> {
-        // `our sub` is package-scoped: a duplicate across sibling blocks is
-        // X::Redeclaration even though each block is its own lexical scope.
-        if let Some(e) = Self::find_our_routine_redeclaration(stmts, &mut HashSet::new()) {
-            return Err(e);
-        }
-        // Type-like declarations (class/role/subset/enum) share one symbol
-        // namespace; a non-stub declaration cannot be redeclared by another
-        // type declaration of the same name (regardless of kind).
-        // `bool` value records whether the existing declaration is a forward stub.
-        let mut seen_types: HashMap<String, bool> = HashMap::new();
-        // Routine declarations (`sub`) share a separate namespace; a routine may
-        // only be redeclared if every declaration of that name is `multi`.
-        let mut seen_routines: HashMap<String, bool> = HashMap::new();
-        // Nested type names declared directly inside each top-level class body,
-        // so that `augment class C { class Nested {} }` can detect redeclaring a
-        // nested type that the original `class C { class Nested {} }` declared.
-        let mut class_nested: HashMap<String, HashSet<String>> = HashMap::new();
-        // The `repr` of each class at its first declaration. A later declaration
-        // (e.g. upgrading a `{ ... }` stub) may not introduce a different repr —
-        // that is X::TooLateForREPR ("must be set at initial declaration").
-        let mut seen_class_repr: HashMap<String, Option<String>> = HashMap::new();
-        let collect_nested = |body: &[Stmt]| -> HashSet<String> {
-            body.iter()
-                .filter_map(|s| match s {
-                    Stmt::ClassDecl { name, .. }
-                    | Stmt::RoleDecl { name, .. }
-                    | Stmt::SubsetDecl { name, .. }
-                    | Stmt::EnumDecl { name, .. } => Some(name.resolve().to_string()),
-                    _ => None,
-                })
-                .collect()
-        };
-        // A forward stub body is a single `... { !!! }` / `{ ??? }` placeholder,
-        // compiled to a lone `__mutsu_stub_die` / `__mutsu_stub_warn` call.
-        let body_is_stub = |body: &[Stmt]| -> bool {
-            let body_no_sl: Vec<_> = body
-                .iter()
-                .filter(|s| !matches!(s, Stmt::SetLine(_)))
-                .collect();
-            body_no_sl.len() == 1
-                && matches!(body_no_sl[0], Stmt::Expr(Expr::Call { name: fn_name, .. })
-                    if *fn_name == "__mutsu_stub_die" || *fn_name == "__mutsu_stub_warn")
-        };
-
-        // Within one class/role/grammar body, two `my`/`our`-scoped methods (or
-        // tokens) of the same name are an X::Redeclaration (a plain `method`/`token`
-        // redeclared is the separate X::Method::Duplicate, not handled here).
-        // Mirrors Rakudo.
-        let dup_scoped_method = |body: &[Stmt]| -> Option<RuntimeError> {
-            let mut seen_my: HashSet<String> = HashSet::new();
-            let mut seen_our: HashSet<String> = HashSet::new();
-            for s in body {
-                let (name, is_my, is_our, what) = match s {
-                    Stmt::MethodDecl {
-                        name,
-                        is_my,
-                        is_our,
-                        ..
-                    } => (name, *is_my, *is_our, "method"),
-                    Stmt::TokenDecl {
-                        name,
-                        is_my,
-                        is_our,
-                        ..
-                    } => (name, *is_my, *is_our, "regex"),
-                    _ => continue,
-                };
-                let n = name.resolve().to_string();
-                if n.is_empty() {
-                    continue;
-                }
-                let dup = (is_my && !seen_my.insert(n.clone()))
-                    || (is_our && !seen_our.insert(n.clone()));
-                if dup {
-                    let mut attrs = std::collections::HashMap::new();
-                    attrs.insert("symbol".to_string(), Value::str(n.clone()));
-                    attrs.insert("what".to_string(), Value::str(what.to_string()));
-                    attrs.insert(
-                        "message".to_string(),
-                        Value::str(format!("Redeclaration of {} '{}'", what, n)),
-                    );
-                    return Some(RuntimeError::typed("X::Redeclaration", attrs));
-                }
-            }
-            None
-        };
-
-        let type_redeclaration = |name: &str| -> RuntimeError {
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("symbol".to_string(), Value::str(name.to_string()));
-            attrs.insert(
-                "message".to_string(),
-                Value::str(format!("Redeclaration of symbol '{}'", name)),
-            );
-            RuntimeError::typed("X::Redeclaration", attrs)
-        };
-
-        // mutsu currently mis-parses `anon sub foo {}` as a bare `anon` term
-        // followed by a normal named SubDecl; an anonymous routine installs no
-        // symbol, so skip the SubDecl that immediately follows a bare `anon`.
-        // TODO: parse `anon sub`/`anon multi` into a genuinely anonymous routine
-        // so this textual heuristic is unnecessary.
-        let mut prev_was_anon = false;
-        for stmt in stmts {
-            // SetLine markers are interleaved between real statements; they must
-            // not reset the "previous statement was `anon`" tracking.
-            if matches!(stmt, Stmt::SetLine(_)) {
-                continue;
-            }
-            let was_anon = prev_was_anon;
-            prev_was_anon = matches!(stmt, Stmt::Expr(Expr::BareWord(w)) if w == "anon");
-
-            // Routine redeclaration (sub a {}; sub a {} / sub a {}; multi sub a {}).
-            if let Stmt::SubDecl { name, multi, .. } = stmt {
-                if was_anon {
-                    continue;
-                }
-                let name = name.resolve().to_string();
-                if name.is_empty() {
-                    continue;
-                }
-                match seen_routines.get(&name) {
-                    None => {
-                        seen_routines.insert(name, *multi);
-                    }
-                    Some(prev_all_multi) => {
-                        if *prev_all_multi && *multi {
-                            // still all-multi, allowed
-                        } else {
-                            let mut attrs = std::collections::HashMap::new();
-                            attrs.insert("symbol".to_string(), Value::str(name.clone()));
-                            attrs.insert("what".to_string(), Value::str("routine".to_string()));
-                            attrs.insert(
-                                "message".to_string(),
-                                Value::str(format!(
-                                    "Redeclaration of routine '{}'. Did you mean to declare a multi-sub?",
-                                    name
-                                )),
-                            );
-                            return Err(RuntimeError::typed("X::Redeclaration", attrs));
-                        }
-                    }
-                }
-                continue;
-            }
-
-            // `augment class C { class Nested {} }` redeclares a nested type that
-            // the original declaration of C already declared.
-            if let Stmt::AugmentClass { name, body, .. } = stmt {
-                let cname = name.resolve().to_string();
-                if let Some(existing) = class_nested.get(&cname) {
-                    for nested in collect_nested(body) {
-                        if existing.contains(&nested) {
-                            return Err(type_redeclaration(&nested));
-                        }
-                    }
-                }
-                continue;
-            }
-
-            // Type-like declarations: class / role / subset / enum.
-            let (name, is_stub) = match stmt {
-                Stmt::ClassDecl {
-                    name,
-                    body,
-                    is_lexical,
-                    repr,
-                    ..
-                } => {
-                    if let Some(err) = dup_scoped_method(body) {
-                        return Err(err);
-                    }
-                    let name = name.resolve().to_string();
-                    // X::TooLateForREPR: the repr must be fixed at the initial
-                    // declaration. If an earlier declaration of this class had a
-                    // different (typically absent) repr, a later one cannot set it.
-                    match seen_class_repr.get(&name) {
-                        Some(prev) if repr.is_some() && prev != repr => {
-                            let mut attrs = std::collections::HashMap::new();
-                            attrs.insert("type".to_string(), Value::str(name.clone()));
-                            attrs.insert(
-                                "message".to_string(),
-                                Value::str(format!(
-                                    "Cannot change REPR of {} now (must be set at initial declaration)",
-                                    name
-                                )),
-                            );
-                            return Err(RuntimeError::typed("X::TooLateForREPR", attrs));
-                        }
-                        Some(_) => {}
-                        None => {
-                            seen_class_repr.insert(name.clone(), repr.clone());
-                        }
-                    }
-                    class_nested
-                        .entry(name.clone())
-                        .or_default()
-                        .extend(collect_nested(body));
-                    let is_stub = body_is_stub(body);
-                    // A non-stub, non-lexical class that already exists in the
-                    // outer environment cannot be redeclared. Lexical classes
-                    // (`my class`) may shadow outer names. Only check
-                    // package-qualified names (containing `::`): simple names may
-                    // have leaked into the global registry from block-scoped
-                    // declarations (a known scoping limitation).
-                    if !is_stub && !*is_lexical && name.contains("::") && self.has_class(&name) {
-                        return Err(type_redeclaration(&name));
-                    }
-                    (name, is_stub)
-                }
-                Stmt::RoleDecl { name, body, .. } => {
-                    if let Some(err) = dup_scoped_method(body) {
-                        return Err(err);
-                    }
-                    (name.resolve().to_string(), body_is_stub(body))
-                }
-                Stmt::SubsetDecl { name, .. } => (name.resolve().to_string(), false),
-                Stmt::EnumDecl { name, variants, .. } => {
-                    // Enum value names share the type/term namespace, so a later
-                    // class/role/enum (or another value) of the same name is a
-                    // redeclaration. Register each value name as a non-stub
-                    // symbol. (`__DYNAMIC__` is the placeholder for an unresolved
-                    // parenthesised body, handled by the undeclared-names check.)
-                    for (vname, _) in variants {
-                        if vname.is_empty() || vname == "__DYNAMIC__" {
-                            continue;
-                        }
-                        if matches!(seen_types.get(vname), Some(false)) {
-                            return Err(type_redeclaration(vname));
-                        }
-                        seen_types.insert(vname.clone(), false);
-                    }
-                    (name.resolve().to_string(), false)
-                }
-                _ => continue,
-            };
-
-            if name.is_empty() {
-                continue;
-            }
-
-            match seen_types.get(&name) {
-                None => {
-                    seen_types.insert(name, is_stub);
-                }
-                Some(false) => {
-                    return Err(type_redeclaration(&name));
-                }
-                Some(true) if is_stub => {
-                    return Err(type_redeclaration(&name));
-                }
-                Some(true) => {
-                    // A forward stub followed by its real definition: upgrade.
-                    seen_types.insert(name, false);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Check for undeclared variable references in EVAL'd code.
-    /// Collects declared variable names from VarDecl/Assign nodes and checks
-    /// Var references against them and the outer environment.
-    /// Placeholder variables ($^x, @_, ...) used directly in the mainline are
-    /// X::Placeholder::Mainline. This must take precedence over the undeclared
-    /// check (otherwise `@_` would be reported as X::Undeclared).
-    pub(crate) fn check_eval_mainline_placeholders(
-        &self,
-        stmts: &[Stmt],
-    ) -> Result<(), RuntimeError> {
-        if let Some(ph) = crate::ast::collect_unattached_placeholders(stmts)
-            .into_iter()
-            .next()
-        {
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("placeholder".to_string(), Value::str(ph.clone()));
-            attrs.insert(
-                "message".to_string(),
-                Value::str(format!(
-                    "Cannot use placeholder parameter {} outside of a sub or block",
-                    ph
-                )),
-            );
-            return Err(RuntimeError::typed("X::Placeholder::Mainline", attrs));
-        }
-        Ok(())
-    }
-
-    pub(crate) fn check_eval_undeclared_vars(&self, stmts: &[Stmt]) -> Result<(), RuntimeError> {
-        let mut declared: HashSet<String> = HashSet::new();
-        for stmt in stmts {
-            Self::check_self_referential_init(stmt)?;
-            Self::collect_declared_vars(stmt, &mut declared);
-            if let Some((sigil, var_name, suggestions)) =
-                self.find_undeclared_var_in_stmt(stmt, &declared)
-            {
-                let symbol = format!("{}{}", sigil, var_name);
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("name".to_string(), Value::str(symbol.clone()));
-                attrs.insert("symbol".to_string(), Value::str(symbol.clone()));
-                // `post` is the source text following the eject point. For a bare
-                // undeclared variable reference, that is the symbol itself.
-                attrs.insert("post".to_string(), Value::str(symbol.clone()));
-                // `highexpect` is the list of additional things the parser was
-                // still expecting at the eject point. For an undeclared variable
-                // there is nothing else expected, so it is an empty list.
-                attrs.insert("highexpect".to_string(), Value::array(vec![]));
-                let mut message = format!("Variable '{}' is not declared.", symbol);
-                if suggestions.len() == 1 {
-                    message.push_str(&format!(" Did you mean '{}'?", suggestions[0]));
-                } else if suggestions.len() > 1 {
-                    let quoted: Vec<String> =
-                        suggestions.iter().map(|s| format!("'{}'", s)).collect();
-                    message.push_str(&format!(
-                        " Did you mean any of these: {}?",
-                        quoted.join(", ")
-                    ));
-                }
-                attrs.insert(
-                    "suggestions".to_string(),
-                    Value::array(suggestions.into_iter().map(Value::str).collect()),
-                );
-                attrs.insert("message".to_string(), Value::str(message));
-                return Err(RuntimeError::typed("X::Undeclared", attrs));
-            }
-        }
-        Ok(())
-    }
-
-    fn check_self_referential_init(stmt: &Stmt) -> Result<(), RuntimeError> {
-        let (name, expr, custom_traits) = match stmt {
-            Stmt::VarDecl {
-                name,
-                expr,
-                custom_traits,
-                ..
-            } => (name, expr, custom_traits),
-            Stmt::SyntheticBlock(body) => {
-                for s in body {
-                    Self::check_self_referential_init(s)?;
-                }
-                return Ok(());
-            }
-            _ => return Ok(()),
-        };
-        let has_init = custom_traits.iter().any(|(n, _)| n == "__has_initializer")
-            || !matches!(expr, Expr::Literal(Value::Nil));
-        let self_ref_expr = has_init && Self::expr_references_var(expr, name);
-        let self_ref_trait = custom_traits.iter().any(|(_, arg)| {
-            arg.as_ref()
-                .is_some_and(|e| Self::expr_references_var(e, name))
-        });
-        if self_ref_expr || self_ref_trait {
-            let sigil = if name.starts_with('@') || name.starts_with('%') {
-                ""
-            } else {
-                "$"
-            };
-            let symbol = format!("{}{}", sigil, name);
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("name".to_string(), Value::str(symbol));
-            return Err(RuntimeError::typed(
-                "X::Syntax::Variable::Initializer",
-                attrs,
-            ));
-        }
-        Ok(())
-    }
-
-    fn expr_references_var(expr: &Expr, var_name: &str) -> bool {
-        let bare = var_name
-            .strip_prefix('%')
-            .or_else(|| var_name.strip_prefix('@'))
-            .or_else(|| var_name.strip_prefix('$'))
-            .unwrap_or(var_name);
-        match expr {
-            Expr::Var(name) => name == var_name || name == bare,
-            Expr::HashVar(name) | Expr::ArrayVar(name) => name == bare,
-            Expr::Binary { left, right, .. } | Expr::MetaOp { left, right, .. } => {
-                Self::expr_references_var(left, var_name)
-                    || Self::expr_references_var(right, var_name)
-            }
-            Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } => {
-                Self::expr_references_var(expr, var_name)
-            }
-            Expr::MethodCall { target, args, .. } | Expr::HyperMethodCall { target, args, .. } => {
-                Self::expr_references_var(target, var_name)
-                    || args.iter().any(|a| Self::expr_references_var(a, var_name))
-            }
-            Expr::Index { target, index, .. } => {
-                Self::expr_references_var(target, var_name)
-                    || Self::expr_references_var(index, var_name)
-            }
-            Expr::ArrayLiteral(items) => {
-                items.iter().any(|i| Self::expr_references_var(i, var_name))
-            }
-            _ => false,
-        }
-    }
-
     /// Suggest lexically-declared variable names close to an undeclared
     /// `symbol` (sigiled, e.g. `$Foo`). Candidate names from `declared` are
     /// normalized to sigiled form (scalars are stored without `$`; arrays/hashes
     /// keep their sigil). Matching is sigil-insensitive on the name part so that
     /// e.g. `$barf` suggests `@barf`. Returns suggestions sorted by edit distance.
-    fn suggest_declared_vars(symbol: &str, declared: &HashSet<String>) -> Vec<String> {
+    pub(super) fn suggest_declared_vars(symbol: &str, declared: &HashSet<String>) -> Vec<String> {
         use crate::runtime::did_you_mean::levenshtein_distance;
         let sigiled = |name: &str| -> String {
             if name.starts_with(['$', '@', '%', '&']) {
@@ -755,1262 +350,4 @@ impl Interpreter {
         scored.sort_by(|a, b| a.0.cmp(&b.0).then_with(|| a.1.cmp(&b.1)));
         scored.into_iter().map(|(_, s)| s).collect()
     }
-
-    fn collect_declared_vars(stmt: &Stmt, out: &mut HashSet<String>) {
-        match stmt {
-            Stmt::VarDecl { name, expr, .. } => {
-                out.insert(name.clone());
-                Self::collect_declared_vars_in_expr(expr, out);
-            }
-            Stmt::Assign { name, .. } => {
-                out.insert(name.clone());
-            }
-            Stmt::SyntheticBlock(body) => {
-                for s in body {
-                    Self::collect_declared_vars(s, out);
-                }
-            }
-            Stmt::For { param, body, .. } => {
-                if let Some(v) = param {
-                    out.insert(v.clone());
-                }
-                for s in body {
-                    Self::collect_declared_vars(s, out);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn collect_declared_vars_in_expr(expr: &Expr, out: &mut HashSet<String>) {
-        if let Expr::DoStmt(stmt) = expr {
-            Self::collect_declared_vars(stmt, out);
-        }
-    }
-
-    /// Find an undeclared Var reference in a statement.
-    /// Returns (sigil, name, suggestions) if undeclared, None otherwise.
-    fn find_undeclared_var_in_stmt(
-        &self,
-        stmt: &Stmt,
-        declared: &HashSet<String>,
-    ) -> Option<(&'static str, String, Vec<String>)> {
-        match stmt {
-            Stmt::Say(exprs) | Stmt::Print(exprs) | Stmt::Note(exprs) | Stmt::Put(exprs) => {
-                for expr in exprs {
-                    if let Expr::BareWord(name) = expr
-                        && !self.has_function(name)
-                        && !self.has_class(name)
-                    {
-                        let suggestions =
-                            Self::suggest_declared_vars(&format!("${}", name), declared);
-                        return Some(("$", name.clone(), suggestions));
-                    }
-                    if let Some(result) = self.find_undeclared_var_in_expr(expr, declared) {
-                        return Some(result);
-                    }
-                }
-                None
-            }
-            Stmt::Expr(expr) => self.find_undeclared_var_in_expr(expr, declared),
-            // Descend into `sub`/`method` bodies so undeclared variables used
-            // inside them (e.g. `sub greet($name) { say "$nam" }`) are caught at
-            // EVAL/compile time. The body's lexical scope adds the routine's
-            // parameters (and, for methods inside a class, the attributes).
-            Stmt::SubDecl { params, body, .. } => {
-                let mut inner = declared.clone();
-                Self::add_routine_locals(params, body, &mut inner);
-                self.find_undeclared_var_in_body(body, &inner)
-            }
-            Stmt::MethodDecl { params, body, .. } => {
-                let mut inner = declared.clone();
-                Self::add_routine_locals(params, body, &mut inner);
-                self.find_undeclared_var_in_body(body, &inner)
-            }
-            Stmt::ClassDecl { body, .. } | Stmt::RoleDecl { body, .. } => {
-                // Attributes become `$!x` / `$.x` (and a bare `$x` only for
-                // `has $x` aliases). Methods in the body see them in scope.
-                let mut inner = declared.clone();
-                Self::add_class_attributes(body, &mut inner);
-                // Class-body lexicals (`my @a`) are also in scope for sibling
-                // statements/methods; collect them all up front.
-                for s in body {
-                    Self::collect_declared_vars(s, &mut inner);
-                }
-                for s in body {
-                    if let Some(result) = self.find_undeclared_var_in_stmt(s, &inner) {
-                        return Some(result);
-                    }
-                }
-                None
-            }
-            _ => None,
-        }
-    }
-
-    /// Seed a lexical scope with a routine's parameters and any variables its
-    /// body declares, so they are not flagged as undeclared.
-    fn add_routine_locals(params: &[String], body: &[Stmt], out: &mut HashSet<String>) {
-        for p in params {
-            out.insert(p.trim_start_matches(['$', '@', '%', '&']).to_string());
-        }
-        for s in body {
-            Self::collect_declared_vars(s, out);
-        }
-    }
-
-    /// Seed a scope with a class/role body's attributes. `has $.name` makes
-    /// `$!name` and `$.name` available (but not a bare `$name`); `has $x`
-    /// (alias form) additionally makes `$x` available.
-    fn add_class_attributes(body: &[Stmt], out: &mut HashSet<String>) {
-        for s in body {
-            if let Stmt::HasDecl { name, is_alias, .. } = s {
-                let attr = name.resolve();
-                // Rakudo's "did you mean" for a bare `$name` that refers to an
-                // attribute suggests `$!name` (the private accessor form).
-                out.insert(format!("!{}", attr));
-                if *is_alias {
-                    out.insert(attr);
-                }
-            }
-        }
-    }
-
-    /// Scan a routine body for the first undeclared variable reference.
-    fn find_undeclared_var_in_body(
-        &self,
-        body: &[Stmt],
-        declared: &HashSet<String>,
-    ) -> Option<(&'static str, String, Vec<String>)> {
-        for s in body {
-            if let Some(result) = self.find_undeclared_var_in_stmt(s, declared) {
-                return Some(result);
-            }
-        }
-        None
-    }
-
-    /// Find an undeclared Var reference in an expression.
-    /// Returns (sigil, name, suggestions) if undeclared, None otherwise.
-    fn find_undeclared_var_in_expr(
-        &self,
-        expr: &Expr,
-        declared: &HashSet<String>,
-    ) -> Option<(&'static str, String, Vec<String>)> {
-        match expr {
-            Expr::Var(name) => {
-                // Skip special variables, variables with twigils, and capture vars ($0, $1, ...)
-                if name == "_"
-                    || name == "/"
-                    || name == "!"
-                    || name.starts_with('*')
-                    || name.starts_with('?')
-                    || name.starts_with('^')
-                    || name.starts_with('~')
-                    || name.starts_with('.')
-                    || name.starts_with('!')
-                    || name.contains("::")
-                    || name.starts_with("CALLER")
-                    || name.starts_with("DYNAMIC")
-                    || name.starts_with("__ANON_")
-                    || name.chars().next().is_some_and(|c| c.is_ascii_digit())
-                {
-                    return None;
-                }
-                // Check if declared locally
-                if declared.contains(name) {
-                    return None;
-                }
-                // Check if in the outer environment (with $ prefix)
-                let sigiled = format!("${}", name);
-                if self.env.contains_key(&sigiled) {
-                    return None;
-                }
-                // A bare-name environment entry only declares the scalar `$name`
-                // if it is an actual variable. A class/sub named `Foo` is stored
-                // bare (as a type object), but does NOT declare the scalar `$Foo`
-                // -- in Raku `$Foo` is its own (undeclared) symbol.
-                if self.env.contains_key(name)
-                    && !self.registry().classes.contains_key(name)
-                    && !self.has_function(name)
-                {
-                    return None;
-                }
-                let suggestions = Self::suggest_declared_vars(&sigiled, declared);
-                Some(("$", name.clone(), suggestions))
-            }
-            Expr::ArrayVar(name) => {
-                if name.starts_with("__ANON_") {
-                    return None;
-                }
-                let sigiled = format!("@{}", name);
-                if declared.contains(name) || declared.contains(&sigiled) {
-                    return None;
-                }
-                if self.env.contains_key(&sigiled) || self.env.contains_key(name) {
-                    return None;
-                }
-                let suggestions = Self::suggest_declared_vars(&sigiled, declared);
-                Some(("@", name.clone(), suggestions))
-            }
-            Expr::HashVar(name) => {
-                if name.starts_with("__ANON_") {
-                    return None;
-                }
-                let sigiled = format!("%{}", name);
-                if declared.contains(name) || declared.contains(&sigiled) {
-                    return None;
-                }
-                if self.env.contains_key(&sigiled) || self.env.contains_key(name) {
-                    return None;
-                }
-                let suggestions = Self::suggest_declared_vars(&sigiled, declared);
-                Some(("%", name.clone(), suggestions))
-            }
-            // Recurse into string interpolation, indexing, and method calls
-            Expr::StringInterpolation(parts) => {
-                for part in parts {
-                    if let Some(result) = self.find_undeclared_var_in_expr(part, declared) {
-                        return Some(result);
-                    }
-                }
-                None
-            }
-            Expr::Index { target, index, .. } => self
-                .find_undeclared_var_in_expr(target, declared)
-                .or_else(|| self.find_undeclared_var_in_expr(index, declared)),
-            Expr::MethodCall { target, args, .. } => {
-                if let Some(result) = self.find_undeclared_var_in_expr(target, declared) {
-                    return Some(result);
-                }
-                for arg in args {
-                    if let Some(result) = self.find_undeclared_var_in_expr(arg, declared) {
-                        return Some(result);
-                    }
-                }
-                None
-            }
-            _ => None,
-        }
-    }
-
-    /// Check for undeclared type names in EVAL'd code.
-    /// Walks the AST looking for BareWord expressions that start with uppercase
-    /// and aren't known types, classes, or packages. This mirrors Raku's
-    /// compile-time check for undeclared symbols.
-    /// A `BEGIN { ... }` block runs at compile time, so it can only see routines
-    /// declared *before* it. Calling a sub that is declared *later* in the same
-    /// unit (`BEGIN { ohnoes() }; sub ohnoes() {}`) is X::Undeclared::Symbols at
-    /// BEGIN time, even though the sub exists by the end of the unit.
-    pub(crate) fn check_eval_begin_forward_calls(
-        &self,
-        stmts: &[Stmt],
-    ) -> Result<(), RuntimeError> {
-        // All sub names declared at this level (forward + backward).
-        let mut all_subs: HashSet<String> = HashSet::new();
-        for s in stmts {
-            if let Stmt::SubDecl { name, .. } = s {
-                all_subs.insert(name.resolve());
-            }
-        }
-        if all_subs.is_empty() {
-            return Ok(());
-        }
-        let mut declared_before: HashSet<String> = HashSet::new();
-        for s in stmts {
-            match s {
-                Stmt::SubDecl { name, .. } => {
-                    declared_before.insert(name.resolve());
-                }
-                Stmt::Phaser {
-                    kind: PhaserKind::Begin,
-                    body,
-                } => {
-                    let mut calls: HashSet<String> = HashSet::new();
-                    Self::collect_call_names_in_stmts(body, &mut calls);
-                    // A call to a sub declared only *after* this BEGIN.
-                    if let Some(fwd) = calls
-                        .iter()
-                        .find(|c| all_subs.contains(*c) && !declared_before.contains(*c))
-                    {
-                        let mut attrs = std::collections::HashMap::new();
-                        attrs.insert("symbol".to_string(), Value::str(fwd.clone()));
-                        attrs.insert(
-                            "message".to_string(),
-                            Value::str(format!("Undeclared routine:\n    {} used at line 1", fwd)),
-                        );
-                        return Err(RuntimeError::typed("X::Undeclared::Symbols", attrs));
-                    }
-                }
-                _ => {}
-            }
-        }
-        Ok(())
-    }
-
-    /// Collect the names of bare function calls (`foo(...)`) appearing anywhere
-    /// in `stmts`, recursing into nested expressions and block bodies.
-    fn collect_call_names_in_stmts(stmts: &[Stmt], out: &mut HashSet<String>) {
-        for s in stmts {
-            match s {
-                Stmt::Expr(e) => Self::collect_call_names_in_expr(e, out),
-                Stmt::VarDecl { expr, .. } | Stmt::Assign { expr, .. } => {
-                    Self::collect_call_names_in_expr(expr, out)
-                }
-                Stmt::Say(es) | Stmt::Print(es) | Stmt::Put(es) | Stmt::Note(es) => {
-                    for e in es {
-                        Self::collect_call_names_in_expr(e, out);
-                    }
-                }
-                Stmt::Block(body) | Stmt::SyntheticBlock(body) => {
-                    Self::collect_call_names_in_stmts(body, out)
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn collect_call_names_in_expr(expr: &Expr, out: &mut HashSet<String>) {
-        match expr {
-            Expr::Call { name, args } => {
-                out.insert(name.resolve());
-                for a in args {
-                    Self::collect_call_names_in_expr(a, out);
-                }
-            }
-            Expr::Binary { left, right, .. } => {
-                Self::collect_call_names_in_expr(left, out);
-                Self::collect_call_names_in_expr(right, out);
-            }
-            Expr::Unary { expr, .. } | Expr::Grouped(expr) => {
-                Self::collect_call_names_in_expr(expr, out)
-            }
-            Expr::MethodCall { target, args, .. } => {
-                Self::collect_call_names_in_expr(target, out);
-                for a in args {
-                    Self::collect_call_names_in_expr(a, out);
-                }
-            }
-            Expr::StringInterpolation(parts) => {
-                for p in parts {
-                    Self::collect_call_names_in_expr(p, out);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    pub(crate) fn check_eval_undeclared_names(&self, stmts: &[Stmt]) -> Result<(), RuntimeError> {
-        // Collect class/role names and locally-declared variables/subs from
-        // this EVAL code. Both are valid references so they should not be
-        // flagged as undeclared bareword names.
-        let mut local_classes: HashSet<String> = HashSet::new();
-        for stmt in stmts {
-            if let Stmt::ClassDecl { name, .. } = stmt {
-                local_classes.insert(name.resolve());
-            }
-            if let Stmt::RoleDecl { name, .. } = stmt {
-                local_classes.insert(name.resolve());
-            }
-        }
-        let mut declared: HashSet<String> = HashSet::new();
-        for stmt in stmts {
-            Self::collect_declared_vars(stmt, &mut declared);
-        }
-        // Normalize collected names: strip leading sigils so bareword lookups
-        // (e.g. `foo` after `my &foo := ...`) resolve correctly.
-        let bare: Vec<String> = declared
-            .iter()
-            .filter_map(|n| n.strip_prefix(['$', '@', '%', '&']).map(|s| s.to_string()))
-            .collect();
-        for n in bare {
-            declared.insert(n);
-        }
-        // Also include any sub/method/grammar/enum names defined at top-level.
-        for stmt in stmts {
-            Self::collect_declared_routine_names(stmt, &mut declared);
-        }
-        for stmt in stmts {
-            if let Some(name) = self.find_undeclared_name_in_stmt(stmt, &local_classes, &declared) {
-                let suggestions = self.suggest_type_names(&name);
-                return Err(RuntimeError::undeclared_type_symbols(
-                    &name,
-                    format!("Undeclared name:\n    {} used at line 1", name),
-                    suggestions,
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn collect_declared_routine_names(stmt: &Stmt, out: &mut HashSet<String>) {
-        match stmt {
-            Stmt::SubDecl { name, .. } | Stmt::MethodDecl { name, .. } => {
-                out.insert(name.resolve());
-            }
-            Stmt::EnumDecl { name, variants, .. } => {
-                out.insert(name.resolve());
-                for (vname, _) in variants {
-                    out.insert(vname.clone());
-                }
-            }
-            Stmt::SyntheticBlock(body) => {
-                for s in body {
-                    Self::collect_declared_routine_names(s, out);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    /// Find an undeclared type name (BareWord starting with uppercase) in a statement.
-    fn find_undeclared_name_in_stmt(
-        &self,
-        stmt: &Stmt,
-        local_classes: &HashSet<String>,
-        declared: &HashSet<String>,
-    ) -> Option<String> {
-        match stmt {
-            Stmt::Expr(expr) => self.find_undeclared_name_in_expr(expr, local_classes, declared),
-            // A `constant NAME = <init>` whose initializer references an undeclared
-            // bareword (`constant foo = bar`) is X::Undeclared::Symbols. Scoped to
-            // `constant` only (custom trait `__constant`) to avoid false positives
-            // on ordinary `my $x = ...` initializers that resolve at runtime.
-            Stmt::VarDecl {
-                expr,
-                custom_traits,
-                ..
-            } if custom_traits.iter().any(|(t, _)| t == "__constant") => {
-                self.find_undeclared_name_in_expr(expr, local_classes, declared)
-            }
-            // `given X { when SomeUndeclaredType {...} }`: a `when` whose condition
-            // is an undeclared (bare uppercase) type name is X::Undeclared.
-            Stmt::Given { body, .. } => body
-                .iter()
-                .find_map(|s| self.find_undeclared_name_in_stmt(s, local_classes, declared)),
-            Stmt::When { cond, .. } => {
-                self.find_undeclared_name_in_expr(cond, local_classes, declared)
-            }
-            // `enum E (Foo, Bar)` with bare *terms* in the parenthesised body
-            // (parsed as a single `__DYNAMIC__` value expression) references
-            // undeclared symbols — `<Foo Bar>` is the autoquoting form. Scan the
-            // value expressions for undeclared barewords.
-            Stmt::EnumDecl { variants, .. } => variants.iter().find_map(|(_, vexpr)| {
-                vexpr
-                    .as_ref()
-                    .and_then(|e| self.find_undeclared_name_in_expr(e, local_classes, declared))
-            }),
-            // `use Module BareWord` — a bare identifier as a positional import
-            // argument is a term reference, not an import symbol (those are
-            // strings / `<...>` / `:tags`). An undeclared one is
-            // X::Undeclared::Symbols (rakudo: "Undeclared name: ...").
-            Stmt::Use { arg: Some(arg), .. } | Stmt::No { arg: Some(arg), .. } => {
-                self.find_undeclared_name_in_expr(arg, local_classes, declared)
-            }
-            _ => None,
-        }
-    }
-
-    /// Find an undeclared bareword name in an expression.
-    /// Checks BareWord nodes that are not known types, classes, functions, or
-    /// declared in the current EVAL scope.
-    fn find_undeclared_name_in_expr(
-        &self,
-        expr: &Expr,
-        local_classes: &HashSet<String>,
-        declared: &HashSet<String>,
-    ) -> Option<String> {
-        match expr {
-            Expr::BareWord(name) => {
-                // Skip well-known constants and special names
-                if matches!(
-                    name.as_str(),
-                    "NaN" | "Inf" | "Empty" | "True" | "False" | "Nil" | "Any" | "Mu"
-                ) {
-                    return None;
-                }
-                // Skip Raku keywords / special syntactic words that are
-                // legitimately parsed as BareWord but should not be treated
-                // as undeclared names.
-                if matches!(
-                    name.as_str(),
-                    "self"
-                        | "given"
-                        | "when"
-                        | "default"
-                        | "if"
-                        | "elsif"
-                        | "else"
-                        | "unless"
-                        | "with"
-                        | "without"
-                        | "orwith"
-                        | "for"
-                        | "while"
-                        | "until"
-                        | "loop"
-                        | "repeat"
-                        | "do"
-                        | "try"
-                        | "anon"
-                        | "my"
-                        | "our"
-                        | "has"
-                        | "state"
-                        | "sub"
-                        | "method"
-                        | "submethod"
-                        | "multi"
-                        | "proto"
-                        | "only"
-                        | "class"
-                        | "role"
-                        | "grammar"
-                        | "token"
-                        | "rule"
-                        | "regex"
-                        | "module"
-                        | "package"
-                        | "enum"
-                        | "subset"
-                        | "constant"
-                        | "return"
-                        | "leave"
-                        | "last"
-                        | "next"
-                        | "redo"
-                        | "succeed"
-                        | "proceed"
-                        | "die"
-                        | "fail"
-                        | "is"
-                        | "does"
-                        | "of"
-                        | "where"
-                        | "but"
-                        | "use"
-                        | "no"
-                        | "need"
-                        | "require"
-                        | "import"
-                        | "lazy"
-                        | "eager"
-                        | "hyper"
-                        | "race"
-                        | "sink"
-                        | "react"
-                        | "supply"
-                        | "whenever"
-                        | "start"
-                        | "gather"
-                        | "take"
-                        | "quietly"
-                        | "now"
-                        | "time"
-                        | "rand"
-                        | "pi"
-                        | "e"
-                        | "tau"
-                        | "i"
-                ) {
-                    return None;
-                }
-                // Skip names with :: (package-qualified)
-                if name.contains("::") {
-                    return None;
-                }
-                // Skip if it's a known type, class, role, enum, function, or env entry
-                if self.has_type(name)
-                    || self.has_class(name)
-                    || self.has_function(name)
-                    || self.has_multi_function(name)
-                    || self.env().contains_key(name)
-                    // `our`-scoped constants/variables installed in the package
-                    // survive in `our_vars` even after their lexical block exits.
-                    || self.get_our_var(name).is_some()
-                    || local_classes.contains(name.as_str())
-                    || declared.contains(name.as_str())
-                    || crate::runtime::Interpreter::is_builtin_type(name)
-                    || crate::runtime::Interpreter::is_implicit_zero_arg_builtin(name)
-                {
-                    return None;
-                }
-                Some(name.clone())
-            }
-            Expr::Binary { left, right, .. } => self
-                .find_undeclared_name_in_expr(left, local_classes, declared)
-                .or_else(|| self.find_undeclared_name_in_expr(right, local_classes, declared)),
-            Expr::Unary { expr, .. } => {
-                self.find_undeclared_name_in_expr(expr, local_classes, declared)
-            }
-            Expr::StringInterpolation(parts) => {
-                for part in parts {
-                    if let Some(name) =
-                        self.find_undeclared_name_in_expr(part, local_classes, declared)
-                    {
-                        return Some(name);
-                    }
-                }
-                None
-            }
-            // A list/array literal (e.g. the parenthesised `(Foo, Bar)` body of an
-            // enum) — scan each element for an undeclared bareword term.
-            Expr::ArrayLiteral(items) => items
-                .iter()
-                .find_map(|e| self.find_undeclared_name_in_expr(e, local_classes, declared)),
-            _ => None,
-        }
-    }
-
-    fn eval_result_is_unresolved_bareword(&self, stmts: &[Stmt], result: &Value) -> bool {
-        let [Stmt::Expr(Expr::BareWord(name))] = stmts else {
-            return false;
-        };
-        matches!(result, Value::Str(s) if s.as_str() == name)
-            && !self.env().contains_key(name)
-            && !self.has_class(name)
-            && !self.has_function(name)
-            && !self.has_multi_function(name)
-            && !matches!(name.as_str(), "NaN" | "Inf" | "Empty")
-    }
-
-    /// Collect operator sub names from the current environment for EVAL pre-registration.
-    /// Only collects circumfix/postcircumfix operators since they require parser support
-    /// to recognize their delimiter syntax. Other operator categories (prefix, postfix,
-    /// infix, term) work through runtime dispatch without parser pre-registration.
-    pub(crate) fn collect_operator_sub_names(&self) -> Vec<String> {
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-        // Include all user-defined operators (infix, prefix, postfix,
-        // circumfix, postcircumfix) so the EVAL parser can recognize them.
-        for key in self.registry().functions.keys() {
-            let key_s = key.resolve();
-            let name = if let Some(pos) = key_s.rfind("::") {
-                &key_s[pos + 2..]
-            } else {
-                key_s.as_str()
-            };
-            if name.starts_with("circumfix:")
-                || name.starts_with("postcircumfix:")
-                || name.starts_with("infix:")
-                || name.starts_with("prefix:")
-                || name.starts_with("postfix:")
-            {
-                seen.insert(name.to_string());
-            }
-        }
-        for key in self.env.keys() {
-            if key.starts_with("circumfix:")
-                || key.starts_with("postcircumfix:")
-                || key.starts_with("infix:")
-                || key.starts_with("prefix:")
-                || key.starts_with("postfix:")
-            {
-                seen.insert(key.resolve());
-            }
-        }
-        // Also include operators imported via `use Module` at runtime. This
-        // captures prefix/infix/postfix operators declared with `is export`
-        // in loaded modules, without exposing non-exported subs.
-        for name in &self.imported_operator_names {
-            seen.insert(name.clone());
-        }
-        let mut names: Vec<String> = seen.into_iter().collect();
-        names.sort();
-        names
-    }
-
-    pub(crate) fn collect_operator_assoc_map(&self) -> HashMap<String, String> {
-        let mut assoc = HashMap::new();
-        for (key, value) in &self.operator_assoc {
-            let name = if let Some(pos) = key.rfind("::") {
-                &key[pos + 2..]
-            } else {
-                key.as_str()
-            };
-            if name.starts_with("infix:<") {
-                assoc.insert(name.to_string(), value.clone());
-            }
-        }
-        assoc
-    }
-
-    pub(crate) fn collect_eval_imported_function_names(&self) -> Vec<String> {
-        const TEST_EXPORTS: &[&str] = &[
-            "ok",
-            "nok",
-            "is",
-            "isnt",
-            "is-deeply",
-            "is-approx",
-            "cmp-ok",
-            "like",
-            "unlike",
-            "isa-ok",
-            "does-ok",
-            "can-ok",
-            "lives-ok",
-            "dies-ok",
-            "eval-lives-ok",
-            "eval-dies-ok",
-            "throws-like",
-            "fails-like",
-            "pass",
-            "flunk",
-            "skip",
-            "skip-rest",
-            "todo",
-            "diag",
-            "plan",
-            "done-testing",
-            "bail-out",
-            "subtest",
-            "use-ok",
-            "force_todo",
-            "force-todo",
-            "tap-ok",
-        ];
-        TEST_EXPORTS
-            .iter()
-            .map(|name| (*name).to_string())
-            .collect()
-    }
-
-    /// Collect user-declared subroutine names from the current runtime so
-    /// EVAL'd code can see them as declared at parse time. This allows
-    /// constructs like `first.uc` (where `first` is a user sub shadowing
-    /// the `first` listop builtin) to parse correctly as `first().uc`.
-    pub(crate) fn collect_eval_user_sub_names(&self) -> Vec<String> {
-        let mut names: Vec<String> = Vec::new();
-        for key in self.registry().functions.keys() {
-            let key_s = key.resolve();
-            let short = if let Some(pos) = key_s.rfind("::") {
-                &key_s[pos + 2..]
-            } else {
-                key_s.as_str()
-            };
-            // Skip empty/meta-named entries. Operator subs are handled by
-            // collect_operator_sub_names.
-            if short.is_empty() || short.contains(':') {
-                continue;
-            }
-            names.push(short.to_string());
-        }
-        names
-    }
-
-    pub(super) fn eval_eval_string(&mut self, code: &str) -> Result<Value, RuntimeError> {
-        let routine_snapshot = self.snapshot_routine_registry();
-        let roles_snapshot = self.registry().roles.clone();
-        let user_declared_roles_snapshot =
-            std::mem::take(&mut self.registry_mut().user_declared_roles);
-        let role_candidates_snapshot = self.registry().role_candidates.clone();
-        let role_type_params_snapshot = self.registry().role_type_params.clone();
-        let role_parents_snapshot = self.registry().role_parents.clone();
-        let role_hides_snapshot = self.registry().role_hides.clone();
-        let classes_snapshot = self.registry().classes.clone();
-        let hidden_classes_snapshot = self.registry().hidden_classes.clone();
-        let hidden_defer_parents_snapshot = self.registry().hidden_defer_parents.clone();
-        let class_composed_roles_snapshot = self.registry().class_composed_roles.clone();
-        let class_role_param_bindings_snapshot = self.registry().class_role_param_bindings.clone();
-        let env_snapshot = self.env.clone();
-        let saved_topic = self.env.get("_").cloned();
-        let trimmed = code.trim();
-        if trimmed == "<>" || trimmed == "<STDIN>" {
-            return Err(RuntimeError::new(
-                "X::Obsolete: The degenerate case <> and old angle forms like <STDIN> are disallowed.",
-            ));
-        }
-        let previous_pod = self.env.get("=pod").cloned();
-        let saved_in_eval = self.env.get("__mutsu_in_eval").cloned();
-        // Record the `&name` code-vars that already exist in the enclosing scope,
-        // so a sub declared inside this EVAL may shadow them without being treated
-        // as a redeclaration (see registration_sub.rs).
-        crate::runtime::registration_sub::push_eval_outer_amp_names(
-            self.env
-                .keys()
-                .map(|k| k.resolve().to_string())
-                .filter(|k| k.starts_with('&')),
-        );
-        self.env
-            .insert("__mutsu_in_eval".to_string(), Value::Bool(true));
-        self.collect_pod_blocks(trimmed);
-        // Collect operator sub names so the parser recognizes them in EVAL context
-        let op_names = self.collect_operator_sub_names();
-        let op_assoc = self.collect_operator_assoc_map();
-        let imported_names = self.collect_eval_imported_function_names();
-        let bracketed_stmt_inner = unwrap_bracketed_statements(trimmed)
-            .filter(|inner| looks_like_bracketed_statement_list(inner));
-        // General case: parse and evaluate as Raku code
-        let mut result = if let Some(inner) = bracketed_stmt_inner {
-            // EVAL q[[ ... ]] can yield one wrapper [] around statement lists.
-            self.parse_and_eval_with_operators(inner, &op_names, &op_assoc, &imported_names)
-                .or_else(|_| {
-                    self.parse_and_eval_with_operators(
-                        trimmed,
-                        &op_names,
-                        &op_assoc,
-                        &imported_names,
-                    )
-                })
-        } else {
-            self.parse_and_eval_with_operators(trimmed, &op_names, &op_assoc, &imported_names)
-        };
-        for warning in crate::parser::take_parse_warnings() {
-            self.write_warn_to_stderr(&warning);
-        }
-        // Fallback: parser still rejects forms like `~< foo bar >`.
-        // Rewrite to an equivalent parenthesized form and try again.
-        if result.is_err()
-            && let Some(rewritten) = rewrite_prefixed_angle_list(trimmed)
-        {
-            result = self.parse_and_eval_with_operators(
-                &rewritten,
-                &op_names,
-                &op_assoc,
-                &imported_names,
-            );
-        }
-        // Accept parenthesized statement lists like `(6;)` in EVAL.
-        if result.is_err()
-            && let Some(inner) = unwrap_parenthesized_statements(trimmed)
-        {
-            result =
-                self.parse_and_eval_with_operators(inner, &op_names, &op_assoc, &imported_names);
-        }
-        // EVAL q[[ ... ]] sometimes carries one outer statement-list bracket pair.
-        if result.is_err()
-            && bracketed_stmt_inner.is_none()
-            && let Some(inner) = unwrap_bracketed_statements(trimmed)
-        {
-            result =
-                self.parse_and_eval_with_operators(inner, &op_names, &op_assoc, &imported_names);
-        }
-        // EVAL should accept routine declarations in snippet context.
-        // If unit-scope parsing rejects a declaration, retry inside an implicit block.
-        if result.is_err()
-            && trimmed.contains("sub ")
-            && let Some(err) = result.as_ref().err()
-            && err.message.contains("X::UnitScope::Invalid")
-        {
-            let wrapped = format!("{{ {}; }}", trimmed);
-            let saved_wrapped_eval = self.env.get("__mutsu_eval_wrapped_decls").cloned();
-            self.env
-                .insert("__mutsu_eval_wrapped_decls".to_string(), Value::Bool(true));
-            result =
-                self.parse_and_eval_with_operators(&wrapped, &op_names, &op_assoc, &imported_names);
-            if let Some(saved) = saved_wrapped_eval {
-                self.env
-                    .insert("__mutsu_eval_wrapped_decls".to_string(), saved);
-            } else {
-                self.env.remove("__mutsu_eval_wrapped_decls");
-            }
-        }
-        if let Some(saved) = previous_pod {
-            self.env.insert("=pod".to_string(), saved);
-        } else {
-            self.env.remove("=pod");
-        }
-        if let Some(saved) = saved_in_eval {
-            self.env.insert("__mutsu_in_eval".to_string(), saved);
-        } else {
-            self.env.remove("__mutsu_in_eval");
-        }
-        self.restore_routine_registry_eval(routine_snapshot);
-        let current_roles = self.registry().roles.clone();
-        let current_role_candidates = self.registry().role_candidates.clone();
-        let current_role_type_params = self.registry().role_type_params.clone();
-        let current_role_parents = self.registry().role_parents.clone();
-        let current_role_hides = self.registry().role_hides.clone();
-        let current_classes = self.registry().classes.clone();
-        let current_hidden_classes = self.registry().hidden_classes.clone();
-        let current_hidden_defer_parents = self.registry().hidden_defer_parents.clone();
-        let current_class_composed_roles = self.registry().class_composed_roles.clone();
-        let current_class_role_param_bindings = self.registry().class_role_param_bindings.clone();
-        let current_env = self.env.clone();
-        let current_type_keys: std::collections::HashSet<String> = self
-            .registry()
-            .roles
-            .keys()
-            .chain(self.registry().classes.keys())
-            .cloned()
-            .collect();
-        let snapshot_type_keys: std::collections::HashSet<String> = roles_snapshot
-            .keys()
-            .chain(classes_snapshot.keys())
-            .cloned()
-            .collect();
-        self.registry_mut().roles = roles_snapshot;
-        self.registry_mut().user_declared_roles = user_declared_roles_snapshot;
-        self.registry_mut().role_candidates = role_candidates_snapshot;
-        self.registry_mut().role_type_params = role_type_params_snapshot;
-        self.registry_mut().role_parents = role_parents_snapshot;
-        self.registry_mut().role_hides = role_hides_snapshot;
-        self.registry_mut().classes = classes_snapshot;
-        self.registry_mut().hidden_classes = hidden_classes_snapshot;
-        self.registry_mut().hidden_defer_parents = hidden_defer_parents_snapshot;
-        self.registry_mut().class_composed_roles = class_composed_roles_snapshot;
-        self.registry_mut().class_role_param_bindings = class_role_param_bindings_snapshot;
-        self.registry_mut().roles.extend(current_roles);
-        // Don't extend user_declared_roles: roles declared in EVAL are EVAL-scoped
-        // and should not affect the parent's redeclaration detection.
-        self.registry_mut()
-            .role_candidates
-            .extend(current_role_candidates);
-        self.registry_mut()
-            .role_type_params
-            .extend(current_role_type_params);
-        self.registry_mut()
-            .role_parents
-            .extend(current_role_parents);
-        self.registry_mut().role_hides.extend(current_role_hides);
-        self.registry_mut().classes.extend(current_classes);
-        self.registry_mut()
-            .hidden_classes
-            .extend(current_hidden_classes);
-        self.registry_mut()
-            .hidden_defer_parents
-            .extend(current_hidden_defer_parents);
-        self.registry_mut()
-            .class_composed_roles
-            .extend(current_class_composed_roles);
-        self.registry_mut()
-            .class_role_param_bindings
-            .extend(current_class_role_param_bindings);
-        for key in current_type_keys.union(&snapshot_type_keys) {
-            if let Some(value) = current_env.get(key).cloned() {
-                self.env.insert(key.clone(), value);
-            } else if let Some(value) = env_snapshot.get(key).cloned() {
-                self.env.insert(key.clone(), value);
-            } else {
-                self.env.remove(key);
-            }
-        }
-        // Lexical code variables created inside EVAL must not leak back into the
-        // caller's environment, or repeated EVALs of the same helper sub will
-        // trip routine redeclaration checks on stale `&name` bindings.
-        let callable_keys: std::collections::HashSet<String> = current_env
-            .keys()
-            .chain(env_snapshot.keys())
-            .filter(|key| key.starts_with("&"))
-            .map(|key| key.resolve())
-            .collect();
-        for key in callable_keys {
-            if let Some(value) = env_snapshot.get(&key).cloned() {
-                self.env.insert(key, value);
-            } else {
-                self.env.remove(&key);
-            }
-        }
-        // Restore $_ so EVAL does not clobber the caller's topic variable
-        if let Some(topic) = saved_topic {
-            self.env.insert("_".to_string(), topic);
-        } else {
-            self.env.remove("_");
-        }
-        crate::runtime::registration_sub::pop_eval_outer_amp_names();
-        result
-    }
-
-    /// Build a CallFrame Instance for the given depth.
-    /// `callsite_line` is the line where `callframe()` was called (from hidden arg).
-    pub(super) fn callframe_value(
-        &self,
-        depth: usize,
-        callsite_line: Option<i64>,
-    ) -> Option<Value> {
-        let file = self
-            .env
-            .get("?FILE")
-            .map(|v| v.to_string_value())
-            .unwrap_or_default();
-
-        if depth == 0 {
-            // Current frame: use current env, current file/line, current code
-            let line = callsite_line.unwrap_or(0);
-            let code = self.current_routine_sub_value();
-            let my_hash = self.build_lexical_hash(&self.env, None);
-            let mut attrs = HashMap::new();
-            attrs.insert("line".to_string(), Value::Int(line));
-            attrs.insert("file".to_string(), Value::str(file));
-            Self::insert_callframe_code_attrs(&mut attrs, &code);
-            attrs.insert("code".to_string(), code);
-            attrs.insert("my".to_string(), my_hash);
-            attrs.insert("inline".to_string(), Value::Bool(false));
-            attrs.insert("__depth".to_string(), Value::Int(0));
-            attrs.insert("annotations".to_string(), self.build_annotations(&attrs));
-            return Some(Value::make_instance(Symbol::intern("CallFrame"), attrs));
-        }
-
-        // depth >= 1: walk up the caller env stack
-        let stack_len = self.callframe_stack.len();
-        if depth > stack_len {
-            return None;
-        }
-        let entry = &self.callframe_stack[stack_len - depth];
-        let code = entry.code.clone().unwrap_or(Value::Nil);
-        let my_hash = self.build_lexical_hash(&entry.env, Some(depth));
-        let mut attrs = HashMap::new();
-        attrs.insert("line".to_string(), Value::Int(entry.line));
-        attrs.insert("file".to_string(), Value::str(entry.file.clone()));
-        Self::insert_callframe_code_attrs(&mut attrs, &code);
-        attrs.insert("code".to_string(), code);
-        attrs.insert("my".to_string(), my_hash);
-        attrs.insert("inline".to_string(), Value::Bool(false));
-        attrs.insert("__depth".to_string(), Value::Int(depth as i64));
-        attrs.insert("annotations".to_string(), self.build_annotations(&attrs));
-        Some(Value::make_instance(Symbol::intern("CallFrame"), attrs))
-    }
-
-    /// Extract subname, package, subtype, and sub attributes from a code value
-    /// and insert them into the CallFrame attributes map.
-    fn insert_callframe_code_attrs(attrs: &mut HashMap<String, Value>, code: &Value) {
-        match code {
-            Value::Sub(sd) => {
-                let name = sd.name.resolve();
-                attrs.insert("subname".to_string(), Value::str(name.to_string()));
-                let pkg = sd.package.resolve();
-                attrs.insert("package".to_string(), Value::str(pkg.to_string()));
-                attrs.insert("subtype".to_string(), Value::str("SubRoutine".to_string()));
-                attrs.insert("sub".to_string(), code.clone());
-            }
-            _ => {
-                attrs.insert("subname".to_string(), Value::str(String::new()));
-                attrs.insert("package".to_string(), Value::str(String::new()));
-                attrs.insert("subtype".to_string(), Value::str(String::new()));
-                attrs.insert("sub".to_string(), Value::Nil);
-            }
-        }
-    }
-
-    fn current_routine_sub_value(&self) -> Value {
-        // Try to find the current routine as a Sub value from the block stack
-        for v in self.block_stack.iter().rev() {
-            if matches!(v, Value::Sub(_)) {
-                return v.clone();
-            }
-        }
-        // Fallback: look at the most recent callframe stack entry for the code
-        if let Some(entry) = self.callframe_stack.last()
-            && let Some(ref code) = entry.code
-        {
-            return code.clone();
-        }
-        Value::Nil
-    }
-
-    fn build_lexical_hash(&self, env: &Env, callframe_depth: Option<usize>) -> Value {
-        let mut hash = HashMap::new();
-        for (k, v) in env.iter() {
-            // Skip internal keys and special variables
-            if k.starts_with("__") || k.starts_with("?") || k.starts_with("*") || k.starts_with("=")
-            {
-                continue;
-            }
-            // Skip type names, enum values, and signal names
-            k.with_str(|s| {
-                if s.chars().next().is_some_and(|c| c.is_uppercase()) {
-                    return;
-                }
-                // Add sigil prefix based on the env key convention:
-                // - Keys starting with '@' or '%' already have sigils (array/hash vars)
-                // - Other keys are scalar variables that need '$' prefix
-                let sigiled = if s.starts_with('@') || s.starts_with('%') || s.starts_with('&') {
-                    s.to_string()
-                } else {
-                    format!("${}", s)
-                };
-                hash.insert(sigiled, v.clone());
-            });
-        }
-        // Store callframe depth so assignments can write back to the caller env
-        if let Some(depth) = callframe_depth {
-            hash.insert("__callframe_depth".to_string(), Value::Int(depth as i64));
-        }
-        Value::hash(hash)
-    }
-
-    pub(crate) fn build_annotations(&self, attrs: &HashMap<String, Value>) -> Value {
-        let mut map = HashMap::new();
-        if let Some(file) = attrs.get("file") {
-            map.insert("file".to_string(), file.clone());
-        }
-        if let Some(line) = attrs.get("line") {
-            map.insert("line".to_string(), line.clone());
-        }
-        Value::hash(map)
-    }
-
-    pub(super) fn seconds_from_value(val: Option<Value>) -> Option<f64> {
-        val.and_then(|v| super::to_float_value(&v))
-    }
-
-    pub(super) fn duration_from_seconds(secs: Option<f64>) -> Duration {
-        let secs = secs.unwrap_or(0.0).max(0.0);
-        Duration::from_secs_f64(secs)
-    }
-
-    pub(super) fn hostname() -> String {
-        env::var("HOSTNAME")
-            .ok()
-            .or_else(|| env::var("COMPUTERNAME").ok())
-            .or_else(|| {
-                Command::new("hostname")
-                    .output()
-                    .ok()
-                    .and_then(|output| String::from_utf8(output.stdout).ok())
-                    .map(|s| s.trim().to_string())
-            })
-            .unwrap_or_else(|| "localhost".to_string())
-    }
-
-    pub(super) fn resolve_host(name: &str) -> Vec<String> {
-        format!("{}:0", name)
-            .to_socket_addrs()
-            .map(|addrs| {
-                addrs
-                    .map(|addr| addr.ip().to_string())
-                    .filter(|ip| !ip.is_empty())
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    pub(super) fn make_os_name_value(name: String, mut addresses: Vec<String>) -> Value {
-        if addresses.is_empty() {
-            addresses.push("127.0.0.1".to_string());
-        }
-        let mut info = HashMap::new();
-        info.insert("name".to_string(), Value::str(name.clone()));
-        info.insert("addr".to_string(), Value::str(addresses[0].clone()));
-        info.insert("aliases".to_string(), Value::array(Vec::new()));
-        let addrs_values = addresses.into_iter().map(Value::str).collect();
-        info.insert("addrs".to_string(), Value::array(addrs_values));
-        Value::hash(info)
-    }
-
-    pub(super) fn get_login_name() -> Option<String> {
-        env::var("LOGNAME")
-            .ok()
-            .or_else(|| env::var("USER").ok())
-            .or_else(|| env::var("USERNAME").ok())
-    }
-
-    pub(super) fn send_signal(pid: i64, signal: i64) -> bool {
-        if pid == 0 {
-            return false;
-        }
-        let pid_str = pid.to_string();
-        #[cfg(unix)]
-        {
-            let mut cmd = Command::new("kill");
-            cmd.arg(format!("-{}", signal));
-            cmd.arg(&pid_str);
-            cmd.status().map(|status| status.success()).unwrap_or(false)
-        }
-        #[cfg(windows)]
-        {
-            let mut cmd = Command::new("taskkill");
-            cmd.args(["/PID", &pid_str, "/F"]);
-            cmd.status().map(|status| status.success()).unwrap_or(false)
-        }
-        #[cfg(not(any(unix, windows)))]
-        {
-            false
-        }
-    }
-}
-
-fn rewrite_prefixed_angle_list(code: &str) -> Option<String> {
-    let (prefix, rest) = if let Some(rest) = code.strip_prefix('~') {
-        ('~', rest)
-    } else if let Some(rest) = code.strip_prefix('+') {
-        ('+', rest)
-    } else if let Some(rest) = code.strip_prefix('?') {
-        ('?', rest)
-    } else {
-        return None;
-    };
-    let inner = rest.trim_start();
-    if !inner.starts_with('<') || !inner.ends_with('>') {
-        return None;
-    }
-    Some(format!("{}({})", prefix, inner))
-}
-
-fn unwrap_parenthesized_statements(code: &str) -> Option<&str> {
-    if !code.starts_with('(') || !code.ends_with(')') {
-        return None;
-    }
-    let mut depth = 0usize;
-    for (i, ch) in code.char_indices() {
-        if ch == '(' {
-            depth += 1;
-        } else if ch == ')' {
-            if depth == 0 {
-                return None;
-            }
-            depth -= 1;
-            if depth == 0 && i + ch.len_utf8() != code.len() {
-                return None;
-            }
-        }
-    }
-    if depth != 0 {
-        return None;
-    }
-    let inner = &code[1..code.len() - 1];
-    // Restrict this fallback to statement-list snippets like `(6;)`.
-    // Plain parenthesized expressions should keep normal parse behavior.
-    if !inner.contains(';') {
-        return None;
-    }
-    Some(inner)
-}
-
-fn unwrap_bracketed_statements(code: &str) -> Option<&str> {
-    if !code.starts_with('[') || !code.ends_with(']') {
-        return None;
-    }
-    let mut depth = 0usize;
-    for (i, ch) in code.char_indices() {
-        if ch == '[' {
-            depth += 1;
-        } else if ch == ']' {
-            if depth == 0 {
-                return None;
-            }
-            depth -= 1;
-            if depth == 0 && i + ch.len_utf8() != code.len() {
-                return None;
-            }
-        }
-    }
-    if depth != 0 {
-        return None;
-    }
-    Some(&code[1..code.len() - 1])
-}
-
-fn looks_like_bracketed_statement_list(inner: &str) -> bool {
-    let trimmed = inner.trim_start();
-    if !inner.contains(';') {
-        return false;
-    }
-    matches!(
-        trimmed.split_whitespace().next(),
-        Some(
-            "my" | "our"
-                | "state"
-                | "sub"
-                | "multi"
-                | "proto"
-                | "class"
-                | "role"
-                | "grammar"
-                | "module"
-                | "unit"
-                | "use"
-                | "need"
-        )
-    )
 }
